requirements.txt
httpx>=0.24.0
websockets>=11.0
pandas>=2.0.0
openpyxl>=3.0.0
PyYAML>=6.0

